pyproject.toml
[tool.poetry]
name = "facturas-afip"
version = "0.1.0"
description = ""
authors = ["gsalomone <gianfranco.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
afip-py = "^1.1.1"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
